refactor(layer): clean up a bit of the layer module
